requirements.txt
# ============================================
# Web-Cross v3.0 Dependencies
# ============================================

# Core HTTP & Parsing
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# CLI Interface
rich>=13.0.0
click>=8.1.0

# Web Server
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3

# Templating & Reporting
jinja2>=3.1.0

# ============================================
# NEW: Async & Performance
# ============================================
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.0.0
asyncio-throttle>=1.0.0

# ============================================
# NEW: Data Validation & Configuration
# ============================================
pydantic>=2.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ============================================
# NEW: Database & State Persistence
# ============================================
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ============================================
# NEW: AI/LLM - Groq + Ollama
# ============================================
groq>=0.4.0
# Ollama: Install separately from https://ollama.com/
# Run: ollama pull llama3.2:3b (or llama3.3)

# ============================================
# NEW: PDF Export
# ============================================
weasyprint>=62.0

# ============================================
# NEW: Retry & Resilience
# ============================================
tenacity>=8.2.0

# ============================================
# NEW: WebSocket Testing
# ============================================
websockets>=12.0

# ============================================
# NEW: DNS for Subdomain Takeover
# ============================================
dnspython>=2.5.0

# ============================================
# NEW: Testing & Quality
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ============================================
# NEW: Development & Linting
# ============================================
ruff>=0.2.0
mypy>=1.8.0
pre-commit>=3.6.0
